Move constant from ssg/oval_object_model/general.py to ssg/constants.py
